+++ just_audio/test/just_audio_test.dart
@@ -1293,7 +1293,7 @@ void runTests() {
   });
 
   test('loadConfiguration', () async {
-    final audioLoadConfiguration = AudioLoadConfiguration(
+    const audioLoadConfiguration = AudioLoadConfiguration(
       darwinLoadControl: DarwinLoadControl(),
       androidLoadControl: AndroidLoadControl(),
       androidLivePlaybackSpeedControl: AndroidLivePlaybackSpeedControl(),