+++ just_audio_platform_interface/lib/just_audio_platform_interface.dart
@@ -166,6 +166,14 @@ abstract class AudioPlayerPlatform {
         "setPreferredPeakBitRate() has not been implemented.");
   }
 
+  /// On iOS and macOS, sets the allowsExternalPlayback option, and does nothing
+  /// on other platforms.
+  Future<SetAllowsExternalPlaybackResponse> setAllowsExternalPlayback(
+      SetAllowsExternalPlaybackRequest request) {
+    throw UnimplementedError(
+        "setAllowsExternalPlayback() has not been implemented.");
+  }
+
   /// Seeks to the given index and position.
   Future<SeekResponse> seek(SeekRequest request) {
     throw UnimplementedError("seek() has not been implemented.");
@@ -723,6 +731,25 @@ class SetPreferredPeakBitRateResponse {
       SetPreferredPeakBitRateResponse();
 }
 
+/// Information communicated to the platform implementation when setting the
+/// automaticallyWaitsToMinimizeStalling option.
+class SetAllowsExternalPlaybackRequest {
+  final bool allowsExternalPlayback;
+
+  SetAllowsExternalPlaybackRequest({required this.allowsExternalPlayback});
+
+  Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
+        'allowsExternalPlayback': allowsExternalPlayback,
+      };
+}
+
+/// Information returned by the platform implementation after setting the
+/// automaticallyWaitsToMinimizeStalling option.
+class SetAllowsExternalPlaybackResponse {
+  static SetAllowsExternalPlaybackResponse fromMap(Map<dynamic, dynamic> map) =>
+      SetAllowsExternalPlaybackResponse();
+}
+
 /// Information communicated to the platform implementation when seeking to a
 /// position and index.
 class SeekRequest {
@@ -1104,6 +1131,58 @@ class AndroidLivePlaybackSpeedControlMessage {
       };
 }
 
+/// Progressive audio source options to be communicated with the platform
+/// implementation.
+class ProgressiveAudioSourceOptionsMessage {
+  final AndroidExtractorOptionsMessage? androidExtractorOptions;
+  final DarwinAssetOptionsMessage? darwinAssetOptions;
+
+  const ProgressiveAudioSourceOptionsMessage({
+    this.androidExtractorOptions,
+    this.darwinAssetOptions,
+  });
+
+  Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
+        'androidExtractorOptions': androidExtractorOptions?.toMap(),
+        'darwinAssetOptions': darwinAssetOptions?.toMap(),
+      };
+}
+
+/// Options for loading audio assets on iOS/macOS to be communicated with the
+/// platform implementation.
+class DarwinAssetOptionsMessage {
+  final bool preferPreciseDurationAndTiming;
+
+  const DarwinAssetOptionsMessage({
+    required this.preferPreciseDurationAndTiming,
+  });
+
+  Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
+        'preferPreciseDurationAndTiming': preferPreciseDurationAndTiming,
+      };
+}
+
+/// Options for extracting media files on Android to be communicated with the
+/// platform implementation.
+class AndroidExtractorOptionsMessage {
+  final bool constantBitrateSeekingEnabled;
+  final bool constantBitrateSeekingAlwaysEnabled;
+  final int mp3Flags;
+
+  const AndroidExtractorOptionsMessage({
+    required this.constantBitrateSeekingEnabled,
+    required this.constantBitrateSeekingAlwaysEnabled,
+    required this.mp3Flags,
+  });
+
+  Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
+        'constantBitrateSeekingEnabled': constantBitrateSeekingEnabled,
+        'constantBitrateSeekingAlwaysEnabled':
+            constantBitrateSeekingAlwaysEnabled,
+        'mp3Flags': mp3Flags,
+      };
+}
+
 /// Information about an audio source to be communicated with the platform
 /// implementation.
 abstract class AudioSourceMessage {
@@ -1120,7 +1199,7 @@ abstract class IndexedAudioSourceMessage extends AudioSourceMessage {
   /// Since the tag type is unknown, this can only be used by platform
   /// implementations that pass by reference.
   final dynamic tag;
-  IndexedAudioSourceMessage({required String id, this.tag}) : super(id: id);
+  IndexedAudioSourceMessage({required super.id, this.tag});
 }
 
 /// Information about a URI audio source to be communicated with the platform
@@ -1130,22 +1209,25 @@ abstract class UriAudioSourceMessage extends IndexedAudioSourceMessage {
   final Map<String, String>? headers;
 
   UriAudioSourceMessage({
-    required String id,
+    required super.id,
     required this.uri,
     this.headers,
-    dynamic tag,
-  }) : super(id: id, tag: tag);
+    super.tag,
+  });
 }
 
 /// Information about a progressive audio source to be communicated with the
 /// platform implementation.
 class ProgressiveAudioSourceMessage extends UriAudioSourceMessage {
+  final ProgressiveAudioSourceOptionsMessage? options;
+
   ProgressiveAudioSourceMessage({
-    required String id,
-    required String uri,
-    Map<String, String>? headers,
-    dynamic tag,
-  }) : super(id: id, uri: uri, headers: headers, tag: tag);
+    required super.id,
+    required super.uri,
+    super.headers,
+    super.tag,
+    this.options,
+  });
 
   @override
   Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
@@ -1153,6 +1235,7 @@ class ProgressiveAudioSourceMessage extends UriAudioSourceMessage {
         'id': id,
         'uri': uri,
         'headers': headers,
+        'options': options?.toMap(),
       };
 }
 
@@ -1160,11 +1243,11 @@ class ProgressiveAudioSourceMessage extends UriAudioSourceMessage {
 /// implementation.
 class DashAudioSourceMessage extends UriAudioSourceMessage {
   DashAudioSourceMessage({
-    required String id,
-    required String uri,
-    Map<String, String>? headers,
-    dynamic tag,
-  }) : super(id: id, uri: uri, headers: headers, tag: tag);
+    required super.id,
+    required super.uri,
+    super.headers,
+    super.tag,
+  });
 
   @override
   Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
@@ -1179,11 +1262,11 @@ class DashAudioSourceMessage extends UriAudioSourceMessage {
 /// implementation.
 class HlsAudioSourceMessage extends UriAudioSourceMessage {
   HlsAudioSourceMessage({
-    required String id,
-    required String uri,
-    Map<String, String>? headers,
-    dynamic tag,
-  }) : super(id: id, uri: uri, headers: headers, tag: tag);
+    required super.id,
+    required super.uri,
+    super.headers,
+    super.tag,
+  });
 
   @override
   Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
@@ -1200,9 +1283,9 @@ class SilenceAudioSourceMessage extends IndexedAudioSourceMessage {
   final Duration duration;
 
   SilenceAudioSourceMessage({
-    required String id,
+    required super.id,
     required this.duration,
-  }) : super(id: id);
+  });
 
   @override
   Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
@@ -1220,11 +1303,11 @@ class ConcatenatingAudioSourceMessage extends AudioSourceMessage {
   final List<int> shuffleOrder;
 
   ConcatenatingAudioSourceMessage({
-    required String id,
+    required super.id,
     required this.children,
     required this.useLazyPreparation,
     required this.shuffleOrder,
-  }) : super(id: id);
+  });
 
   @override
   Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
@@ -1244,12 +1327,12 @@ class ClippingAudioSourceMessage extends IndexedAudioSourceMessage {
   final Duration? end;
 
   ClippingAudioSourceMessage({
-    required String id,
+    required super.id,
     required this.child,
     this.start,
     this.end,
-    dynamic tag,
-  }) : super(id: id, tag: tag);
+    super.tag,
+  });
 
   @override
   Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
@@ -1268,10 +1351,10 @@ class LoopingAudioSourceMessage extends AudioSourceMessage {
   final int count;
 
   LoopingAudioSourceMessage({
-    required String id,
+    required super.id,
     required this.child,
     required this.count,
-  }) : super(id: id);
+  });
 
   @override
   Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
@@ -1395,9 +1478,9 @@ class AndroidLoudnessEnhancerMessage extends AudioEffectMessage {
   final double targetGain;
 
   AndroidLoudnessEnhancerMessage({
-    required bool enabled,
+    required super.enabled,
     required this.targetGain,
-  }) : super(enabled: enabled);
+  });
 
   @override
   Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{
@@ -1487,9 +1570,9 @@ class AndroidEqualizerMessage extends AudioEffectMessage {
   final AndroidEqualizerParametersMessage? parameters;
 
   AndroidEqualizerMessage({
-    required bool enabled,
+    required super.enabled,
     required this.parameters,
-  }) : super(enabled: enabled);
+  });
 
   @override
   Map<dynamic, dynamic> toMap() => <dynamic, dynamic>{