+++ just_audio/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 * 
 
+## 0.9.36
+
+* Add setAllowsExternalPlayback on iOS/macOS.
+* Support index-based seeking on Android/iOS/macOS.
+* Add option to send headers/userAgent without proxy.
+
 ## 0.9.35
 
 * Fix nullable completer argument type (@srawlins).